twitter-stream-service/requirements.txt
async-timeout==4.0.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
greenlet==1.1.2
idna==3.3
oauthlib==3.2.0
packaging==21.3
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.20.0
redis==4.3.1
requests==2.27.1
requests-oauthlib==1.3.1
rq==1.10.1
six==1.16.0
SQLAlchemy==1.4.36
tweepy==4.10.0
urllib3==1.26.9
wrapt==1.14.1

